refactored files and split games and tournament; fix comment 2 4 5 6 8 9 13;
