+++ backend/games/services/tournament.py
@@ -1,94 +1,9 @@
-from .models import Game, Tournament, TournamentPlayer
+from games.models import Game, Tournament, TournamentPlayer
 from users.models import UserProfile
 from django.db import transaction
 from django.shortcuts import get_object_or_404
 from django.db.models import Q, Sum, Case, When, F
 
-class GameService:
-	@staticmethod
-	@transaction.atomic
-	def join_game(game_id: int, user: UserProfile) -> Game:
-		game = get_object_or_404(Game, id=game_id)
-		if game.status != 'pending' or game.player2 or game.player1 == user:
-			raise ValueError('Game cannot be joined')
-		game.player2 = user
-		game.status = 'ready'
-		game.save()
-		return game
-	
-	@staticmethod
-	@transaction.atomic
-	def cancel_game(game_id: int, user: UserProfile) -> Game:
-		game = get_object_or_404(Game, id=game_id)
-		if game.status != 'pending' or game.player2 or game.player1 != user:
-			raise ValueError('Game cannot be canceled')
-		game.status = 'canceled'
-		game.save()
-		return game
-
-	@staticmethod
-	@transaction.atomic
-	def start_game(game_id: int, user: UserProfile) -> Game:
-		game = get_object_or_404(Game, id=game_id)
-		if game.status != 'ready' or (game.player1 != user and game.player2 != user):
-			raise ValueError('Game cannot be started')
-		game.status = 'in_progress'
-		game.save()
-		return game
-	
-	@staticmethod
-	@transaction.atomic
-	def interrupt_game(game_id: int) -> Game:
-		game = get_object_or_404(Game, id=game_id)
-		if game.status != 'in_progress':
-			raise ValueError('Game cannot be interrupted')
-		game.status = 'interrupted'
-		game.save()
-		return game
-	
-	@staticmethod
-	@transaction.atomic
-	def update_game(game_id: int, new_score_player1: int, new_score_player2: int) -> Game:
-		game = get_object_or_404(Game, id=game_id)
-		if game.status != 'in_progress':
-			raise ValueError('Game can not be updated')
-		game.score_player1 = new_score_player1
-		game.score_player2 = new_score_player2
-		if game.score_player1 >= game.winning_score:
-			game.winner = game.player1
-			game.status = 'completed'
-		elif game.score_player2 >= game.winning_score:
-			game.winner = game.player2
-			game.status = 'completed'
-		game.save()
-		tournament = game.tournament
-		if tournament:
-			try:
-				TournamentService.finish_tournament(tournament.id)
-			except ValueError:
-				pass
-		return game
-
-	@staticmethod
-	def calculate_user_statistics(user: UserProfile):
-		completed_games = Game.objects.filter(status='completed').filter(Q(player1=user) | Q(player2=user))
-		total_games = completed_games.count()
-		games_won = completed_games.filter(winner=user).count()
-		games_lost = total_games - games_won
-		win_percentage = (games_won / total_games) * 100 if total_games > 0 else 0
-		total_points = sum(
-			game.score_player1 if game.player1 == user else game.score_player2
-			for game in completed_games
-		)
-		avg_points_per_game = total_points / total_games if total_games > 0 else 0
-
-		return {
-			'total_games': total_games,
-			'games_won': games_won,
-			'games_lost': games_lost,
-			'win_percentage': round(win_percentage, 2),
-			'avgerage_points_per_game': round(avg_points_per_game, 2),
-		}
 
 class TournamentService:
 	@staticmethod
@@ -142,19 +57,21 @@ class TournamentService:
 			raise ValueError('Tournament cannot be started')
 		if user != tournament.creator:
 			raise PermissionError('Only the creator can start the tournament.')
-		players = list(tournament.players.all())
+		players = list(tournament.players.select_related('player').all())
 		if len(players) < tournament.min_players:
-			raise ValueError(f'Tournament cannot be started with less than {tournament.min_players}players')
+			raise ValueError(f'Tournament cannot be started with less than {tournament.min_players} players')
 		
+		games_to_create = []
 		for i in range(len(players)):
 			for j in range(i + 1, len(players)):
-				Game.objects.create(
+				games_to_create.append(Game(
 					tournament=tournament,
 					player1=players[i].player,
 					player2=players[j].player,
 					winning_score=tournament.winning_score,
 					status='ready',
-				)
+				))
+		Game.objects.bulk_create(games_to_create)
 		tournament.status = 'in_progress'
 		tournament.save()
 		return tournament
@@ -219,7 +136,7 @@ class TournamentService:
 		tournament = get_object_or_404(Tournament, id=tournament_id)
 		if tournament.status not in ['in_progress', 'completed']:
 			raise ValueError("Leaderboard only for a in-progress or completed tournament")
-		tournament_players = TournamentPlayer.objects.filter(tournament=tournament)
+		tournament_players = TournamentPlayer.objects.filter(tournament=tournament).select_related('player')
 		leaderboard_data = {}
 		player_stats = []
 		for t_player in tournament_players:
@@ -227,7 +144,7 @@ class TournamentService:
 				Q(player1=t_player.player) | Q(player2=t_player.player),
 				status='completed',
 				tournament=tournament
-			)
+			).select_related('player1', 'player2')
 
 			game_total = completed_games.count()
 			game_win = completed_games.filter(winner=t_player.player).count()
@@ -246,6 +163,7 @@ class TournamentService:
 					'total_score': total_score,
 				}
 			})
+
 		sorted_stats = sorted(
 			player_stats,
 			key=lambda x: (-x['stats']['game_win'], -x['stats']['total_score'])
