+++ backend/games/views/games.py
@@ -7,11 +7,9 @@ from rest_framework.views import APIView
 from django.shortcuts import get_object_or_404
 from django.db.models import Q
 
-from .serializers import GameCreateSerializer, GameDetailSerializer, GameActionSerializer, \
-							TournamentCreateSerializer, TournamentJoinSerializer, \
-							TournamentActionSerializer, TournamentDetailSerializer
-from .models import Game, Tournament
-from .services import GameService, TournamentService
+from games.serializers.games import GameCreateSerializer, GameDetailSerializer, GameActionSerializer
+from games.models import Game
+from games.services import GameService
 
 class GameCreateView(APIView):
 	permission_classes = [IsAuthenticated]
@@ -159,129 +157,3 @@ class GameStatisticsView(APIView):
 	def get(self, request: Request) -> Response:
 		stats = GameService.calculate_user_statistics(request.user)
 		return Response(stats, status=status.HTTP_200_OK)
-
-
-class TournamentCreateView(APIView):
-	permission_classes = [IsAuthenticated]
-
-	@extend_schema(
-		summary="Propose a new tournament",
-		request=TournamentCreateSerializer,
-		tags=['Tournaments'],
-	)
-	def post(self, request: Request) -> Response:
-		serializer = TournamentCreateSerializer(data=request.data)
-		if not serializer.is_valid():
-			return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-		name = serializer.validated_data['name']
-		alias = serializer.validated_data.get('alias', None)
-		user = request.user
-		max_players = serializer.validated_data.get('max_players', 4)
-		min_players = serializer.validated_data.get('min_players', 8)
-		winning_score = serializer.validated_data.get('winnint_score', 10)
-		try:
-			tournament = TournamentService.create_tournament(name, alias, user, max_players, min_players, winning_score)
-			return Response(
-				{
-					'message': 'Tournament created successfully',
-					'status': tournament.status,
-				}, status=status.HTTP_201_CREATED)
-		except ValueError as e:
-			return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
-class TournamentJoinView(APIView):
-	permission_classes = [IsAuthenticated]
-
-	@extend_schema(
-		summary="Join an existing tournament",
-		request=TournamentJoinSerializer,
-		tags=['Tournaments'],
-	)
-	def post(self, request: Request) -> Response:
-		serializer = TournamentJoinSerializer(data=request.data)
-		if not serializer.is_valid():
-			return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-		tournament_id = serializer.validated_data['tournament_id']
-		alias = serializer.validated_data.get('alias', None)
-		user = request.user
-		try:
-			tournament_player = TournamentService.join_tournament(tournament_id, alias, user)
-			return Response(
-				{
-					'message': f'{alias} joined Tournament {tournament_player.tournament.name}',
-					'status': tournament_player.tournament.status,
-				}, status=status.HTTP_200_OK
-			)
-		except ValueError as e:
-			return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
-class TournamentStartView(APIView):
-	permission_classes = [IsAuthenticated]
-
-	@extend_schema(
-		summary="Start a tournament",
-		request=TournamentActionSerializer,
-		tags=['Tournaments'],
-	)
-	def post(self, request: Request) -> Response:
-		serializer = TournamentActionSerializer(data=request.data)
-		if not serializer.is_valid():
-			return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-		tournament_id = serializer.validated_data['tournament_id']
-		try:
-			tournament = TournamentService.start_tournament(tournament_id, request.user)
-			return Response({'message': f'Tournament {tournament.name} started'}, status=status.HTTP_200_OK)
-		except ValueError as e:
-			return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
-class TournamentCancelView(APIView):
-	permission_classes = [IsAuthenticated]
-
-	@extend_schema(
-		summary="Cancel a tournament",
-		request=TournamentActionSerializer,
-		tags=['Tournaments'],
-	)
-	def patch(self, request: Request) -> Response:
-		serializer = TournamentActionSerializer(data=request.data)
-		if not serializer.is_valid():
-			return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-		tournament_id = serializer.validated_data['tournament_id']
-		try:
-			tournament = TournamentService.cancel_tournament(tournament_id, request.user)
-			return Response({'message': f'Tournament {tournament.name} is canceled'}, status=status.HTTP_200_OK)
-		except ValueError as e:
-			return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
-class RegistrationTournamentListView(APIView):
-	permission_classes = [IsAuthenticated]
-
-	@extend_schema(
-		summary="List all the pending games",
-		request=TournamentDetailSerializer,
-		responses={200: TournamentDetailSerializer(many=True)},
-		tags=['Tournaments'],
-	)
-	def get(self, request: Request) -> Response:
-		tournaments = Tournament.objects.filter(status='registration')
-		serializer = TournamentDetailSerializer(tournaments, many=True)
-		return Response(serializer.data, status=status.HTTP_200_OK)
-
-class TournamentLeaderboardView(APIView):
-	permission_classes = [IsAuthenticated]
-
-	@extend_schema(
-		parameters=[OpenApiParameter(name='tournament_id', required=True, type=int)],
-		summary="Show the current situation of tournament",
-		tags=['Tournaments'],
-	)
-	def get(self, request: Request):
-		tournament_id = request.query_params.get('tournament_id')
-		if not tournament_id:
-			return Response({'error': 'Tournament ID is required'}, status=status.HTTP_400_BAD_REQUEST)
-		try:
-			leaderboard = TournamentService.calculate_leaderboard(tournament_id)
-			return Response(leaderboard, status=status.HTTP_200_OK)
-		except ValueError as e:
-			return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
